PWA UI tests refactoring work
