+++ app/src/androidTest/java/org/mozilla/fenix/ui/PwaTest.kt
@@ -11,14 +11,17 @@ import org.mozilla.fenix.helpers.Constants.PackageName.PHONE_APP
 import org.mozilla.fenix.helpers.FeatureSettingsHelper
 import org.mozilla.fenix.helpers.HomeActivityIntentTestRule
 import org.mozilla.fenix.helpers.TestHelper.assertNativeAppOpens
+import org.mozilla.fenix.helpers.TestHelper.openAppFromExternalLink
+import org.mozilla.fenix.ui.robots.addToHomeScreen
+import org.mozilla.fenix.ui.robots.browserScreen
 import org.mozilla.fenix.ui.robots.customTabScreen
+import org.mozilla.fenix.ui.robots.mDevice
 import org.mozilla.fenix.ui.robots.navigationToolbar
 import org.mozilla.fenix.ui.robots.pwaScreen
 
 class PwaTest {
     private val featureSettingsHelper = FeatureSettingsHelper()
-    private val externalLinksPWAPage = "https://mozilla-mobile.github.io/testapp/externalLinks.html"
-    private val clockAndCalendarPWAPage = "https://andiaj.github.io/testapp/clockAndCalendarForm.html"
+    private val htmlControlsPWAPage = "https://andiaj.github.io/testapp/htmlControlsForm.html"
     private val emailLink = "mailto://example@example.com"
     private val phoneLink = "tel://1234567890"
     private val shortcutTitle = "TEST_APP"
@@ -41,20 +44,20 @@ class PwaTest {
     @SmokeTest
     @Test
     fun externalLinkPWATest() {
-        val customTabTitle = "Google"
+        val searchTerm = "Mozilla"
 
         navigationToolbar {
-        }.enterURLAndEnterToBrowser(externalLinksPWAPage.toUri()) {
+        }.enterURLAndEnterToBrowser(htmlControlsPWAPage.toUri()) {
         }.openThreeDotMenu {
         }.clickInstall {
             clickAddAutomaticallyButton()
         }.openHomeScreenShortcut(shortcutTitle) {
             clickLinkMatchingText("External link")
-            fillAndSubmitGoogleSearchQuery("Mozilla")
+            fillAndSubmitDuckDuckGoSearchQuery(searchTerm)
         }
 
         customTabScreen {
-            verifyCustomTabToolbarTitle(customTabTitle)
+            verifyCustomTabToolbarTitle(searchTerm)
         }
     }
 
@@ -63,7 +66,7 @@ class PwaTest {
     fun emailLinkPWATest() {
 
         navigationToolbar {
-        }.enterURLAndEnterToBrowser(externalLinksPWAPage.toUri()) {
+        }.enterURLAndEnterToBrowser(htmlControlsPWAPage.toUri()) {
         }.openThreeDotMenu {
         }.clickInstall {
             clickAddAutomaticallyButton()
@@ -78,7 +81,7 @@ class PwaTest {
     fun telephoneLinkPWATest() {
 
         navigationToolbar {
-        }.enterURLAndEnterToBrowser(externalLinksPWAPage.toUri()) {
+        }.enterURLAndEnterToBrowser(htmlControlsPWAPage.toUri()) {
         }.openThreeDotMenu {
         }.clickInstall {
             clickAddAutomaticallyButton()
@@ -90,10 +93,44 @@ class PwaTest {
 
     @SmokeTest
     @Test
+    fun saveLoginsInPWATest() {
+
+        navigationToolbar {
+        }.enterURLAndEnterToBrowser(htmlControlsPWAPage.toUri()) {
+            verifyNotificationDotOnMainMenu()
+        }.openThreeDotMenu {
+        }.clickInstall {
+            clickAddAutomaticallyButton()
+        }.openHomeScreenShortcut(shortcutTitle) {
+            mDevice.waitForIdle()
+            fillAndSubmitLoginCredentials("mozilla", "firefox")
+            verifySaveLoginPromptIsDisplayed()
+            saveLoginFromPrompt("Save")
+            openAppFromExternalLink(htmlControlsPWAPage)
+
+            browserScreen {
+            }.openThreeDotMenu {
+            }.openSettings {
+            }.openLoginsAndPasswordSubMenu {
+            }.openSavedLogins {
+                verifySecurityPromptForLogins()
+                tapSetupLater()
+                verifySavedLoginFromPrompt("mozilla")
+            }
+
+            addToHomeScreen {
+            }.searchAndOpenHomeScreenShortcut(shortcutTitle) {
+                verifyPrefilledLoginCredentials("mozilla", shortcutTitle)
+            }
+        }
+    }
+
+    @SmokeTest
+    @Test
     fun cancelCalendarFormPWATest() {
 
         navigationToolbar {
-        }.enterURLAndEnterToBrowser(clockAndCalendarPWAPage.toUri()) {
+        }.enterURLAndEnterToBrowser(htmlControlsPWAPage.toUri()) {
         }.openThreeDotMenu {
         }.clickInstall {
             clickAddAutomaticallyButton()
@@ -112,7 +149,7 @@ class PwaTest {
     fun setAndClearCalendarFormPWATest() {
 
         navigationToolbar {
-        }.enterURLAndEnterToBrowser(clockAndCalendarPWAPage.toUri()) {
+        }.enterURLAndEnterToBrowser(htmlControlsPWAPage.toUri()) {
         }.openThreeDotMenu {
         }.clickInstall {
             clickAddAutomaticallyButton()
@@ -136,7 +173,7 @@ class PwaTest {
     fun cancelClockFormPWATest() {
 
         navigationToolbar {
-        }.enterURLAndEnterToBrowser(clockAndCalendarPWAPage.toUri()) {
+        }.enterURLAndEnterToBrowser(htmlControlsPWAPage.toUri()) {
         }.openThreeDotMenu {
         }.clickInstall {
             clickAddAutomaticallyButton()
@@ -155,7 +192,7 @@ class PwaTest {
     fun setAndClearClockFormPWATest() {
 
         navigationToolbar {
-        }.enterURLAndEnterToBrowser(clockAndCalendarPWAPage.toUri()) {
+        }.enterURLAndEnterToBrowser(htmlControlsPWAPage.toUri()) {
         }.openThreeDotMenu {
         }.clickInstall {
             clickAddAutomaticallyButton()
